+++ CHANGELOG.md
@@ -17,6 +17,7 @@ For a full diff see [`e66cfff...main`][e66cfff...main].
 - Added `Location\ClassMethodLocation` ([#5]), by [@localheinz]
 - Added `Location\ClassMethodParameterLocation` ([#6]), by [@localheinz]
 - Added `Location\FunctionLocation` ([#7]), by [@localheinz]
+- Added `Location\FunctionParameterLocation` ([#8]), by [@localheinz]
 
 [e66cfff...main]: https://github.com/ergebnis/attribute-collector/compare/e66cfff...main
 
@@ -27,5 +28,6 @@ For a full diff see [`e66cfff...main`][e66cfff...main].
 [#5]: https://github.com/ergebnis/attribute-collector/pull/5
 [#6]: https://github.com/ergebnis/attribute-collector/pull/6
 [#7]: https://github.com/ergebnis/attribute-collector/pull/7
+[#8]: https://github.com/ergebnis/attribute-collector/pull/8
 
 [@localheinz]: https://github.com/localheinz